refactor(websocket): extract the runner's pure helpers into session modules
T1 of docs/websocket-runner-refactor-plan.md. `unified-websocket-runner.ts`
opened with ~1,400 lines that never touch `this`, so nothing about the file
said which of them belonged to the socket host and which were free functions
the later extractions would need. Moving them out first shrinks the class
file before any behavior-bearing step and gives T3–T6 modules to import from
instead of reaching back into the runner.

Pure move: no renames, no signature changes, no logic edits. Every name the
runner exported before it still exports, so no caller and no test changes.

The auto-save path took its `fetchExternalMedia` call with it, so the URL
egress inventory gains an entry for its new file and the runner's entry now
names only the chat-source-image site it still owns.
